+++ goalc/compiler/IR.cpp
@@ -13,7 +13,7 @@ Register get_reg(const RegVal* rv, const AllocationResult& allocs, emitter::IR_R
     auto reg = rv->rlet_constraint().value();
     if (rv->ireg().id < int(range.size())) {
       auto& lr = range.at(rv->ireg().id);
-      if (irec.ir_id >= lr.min && irec.ir_id <= lr.max) {
+      if (lr.has_info_at(irec.ir_id)) {
         auto ass_reg = range.at(rv->ireg().id).get(irec.ir_id);
         if (ass_reg.kind == Assignment::Kind::REGISTER) {
           assert(ass_reg.reg == reg);
@@ -41,7 +41,7 @@ int get_stack_offset(const RegVal* rv, const AllocationResult& allocs) {
   } else {
     assert(rv->forced_on_stack());
     auto& ass = allocs.ass_as_ranges.at(rv->ireg().id);
-    auto stack_slot = allocs.get_slot_for_spill(ass.assignment.at(0).stack_slot);
+    auto stack_slot = allocs.get_slot_for_spill(ass.stack_slot());
     assert(stack_slot >= 0);
     return stack_slot * 8;
   }
@@ -103,6 +103,7 @@ void regset_common(emitter::ObjectGenerator* gen,
   if (src_class == RegClass::GPR_64 && dst_class == RegClass::GPR_64) {
     if (src_reg == dst_reg) {
       // eliminate move
+      gen->count_eliminated_move();
       gen->add_instr(IGen::null(), irec);
     } else {
       gen->add_instr(IGen::mov_gpr64_gpr64(dst_reg, src_reg), irec);
@@ -110,6 +111,7 @@ void regset_common(emitter::ObjectGenerator* gen,
   } else if (src_class == RegClass::FLOAT && dst_class == RegClass::FLOAT) {
     if (src_reg == dst_reg) {
       // eliminate move
+      gen->count_eliminated_move();
       gen->add_instr(IGen::null(), irec);
     } else {
       gen->add_instr(IGen::mov_xmm32_xmm32(dst_reg, src_reg), irec);
@@ -117,6 +119,7 @@ void regset_common(emitter::ObjectGenerator* gen,
   } else if (src_is_xmm128 && dst_is_xmm128) {
     if (src_reg == dst_reg) {
       // eliminate move
+      gen->count_eliminated_move();
       gen->add_instr(IGen::null(), irec);
     } else {
       gen->add_instr(IGen::mov_vf_vf(dst_reg, src_reg), irec);