+++ goalc/regalloc/Allocator.cpp
@@ -3,87 +3,17 @@
  * Implementation of register allocation algorithms
  */
 
-#include <algorithm>
+#include <stdexcept>
 #include "third-party/fmt/core.h"
 #include "Allocator.h"
+#include "goalc/regalloc/allocator_interface.h"
 
-/*!
- * Find basic blocks and add block link info.
- */
-void find_basic_blocks(RegAllocCache* cache, const AllocationInput& in) {
-  std::vector<int> dividers;
-
-  dividers.push_back(0);
-  dividers.push_back(in.instructions.size());
-
-  // loop over instructions, finding jump targets
-  for (uint32_t i = 0; i < in.instructions.size(); i++) {
-    const auto& instr = in.instructions[i];
-    if (!instr.jumps.empty()) {
-      dividers.push_back(i + 1);
-      for (auto dest : instr.jumps) {
-        dividers.push_back(dest);
-      }
-    }
+std::string LiveInfo::print_assignment() {
+  std::string result = "Assignment for var " + std::to_string(var) + "\n";
+  for (uint32_t i = 0; i < assignment.size(); i++) {
+    result += fmt::format("i[{:3d}] {}\n", i + min, assignment.at(i).to_string());
   }
-
-  // sort dividers, and make blocks
-  std::sort(dividers.begin(), dividers.end(), [](int a, int b) { return a < b; });
-
-  for (uint32_t i = 0; i < dividers.size() - 1; i++) {
-    if (dividers[i] != dividers[i + 1]) {
-      // new basic block!
-      RegAllocBasicBlock block;
-      for (int j = dividers[i]; j < dividers[i + 1]; j++) {
-        block.instr_idx.push_back(j);
-      }
-
-      block.idx = cache->basic_blocks.size();
-      cache->basic_blocks.push_back(block);
-    }
-  }
-
-  if (!cache->basic_blocks.empty()) {
-    cache->basic_blocks.front().is_entry = true;
-    cache->basic_blocks.back().is_exit = true;
-  }
-
-  auto find_basic_block_to_target = [&](int instr) {
-    bool found = false;
-    uint32_t result = -1;
-    for (uint32_t i = 0; i < cache->basic_blocks.size(); i++) {
-      if (!cache->basic_blocks[i].instr_idx.empty() &&
-          cache->basic_blocks[i].instr_idx.front() == instr) {
-        assert(!found);
-        found = true;
-        result = i;
-      }
-    }
-    if (!found) {
-      printf("[RegAlloc Error] couldn't find basic block beginning with instr %d of %d\n", instr,
-             int(in.instructions.size()));
-    }
-    assert(found);
-    return result;
-  };
-
-  // link blocks
-  for (auto& block : cache->basic_blocks) {
-    assert(!block.instr_idx.empty());
-    auto& last_instr = in.instructions.at(block.instr_idx.back());
-    if (last_instr.fallthrough) {
-      // try to link to next block:
-      int next_idx = block.idx + 1;
-      if (next_idx < (int)cache->basic_blocks.size()) {
-        cache->basic_blocks.at(next_idx).pred.push_back(block.idx);
-        block.succ.push_back(next_idx);
-      }
-    }
-    for (auto target : last_instr.jumps) {
-      cache->basic_blocks.at(find_basic_block_to_target(target)).pred.push_back(block.idx);
-      block.succ.push_back(find_basic_block_to_target(target));
-    }
-  }
+  return result;
 }
 
 namespace {
@@ -96,7 +26,7 @@ void compute_live_ranges(RegAllocCache* cache, const AllocationInput& in) {
   cache->live_ranges.resize(cache->max_var, LiveInfo(in.instructions.size(), 0));
 
   // now compute the ranges
-  for (auto& block : cache->basic_blocks) {
+  for (auto& block : cache->control_flow.basic_blocks) {
     // from var use
     for (auto instr_id : block.instr_idx) {
       auto& inst = in.instructions.at(instr_id);
@@ -147,7 +77,7 @@ void analyze_liveliness(RegAllocCache* cache, const AllocationInput& in) {
   }
 
   // phase 1
-  for (auto& block : cache->basic_blocks) {
+  for (auto& block : cache->control_flow.basic_blocks) {
     block.live.resize(block.instr_idx.size());
     block.dead.resize(block.instr_idx.size());
     block.analyze_liveliness_phase1(in.instructions);
@@ -157,16 +87,16 @@ void analyze_liveliness(RegAllocCache* cache, const AllocationInput& in) {
   bool changed = false;
   do {
     changed = false;
-    for (auto& block : cache->basic_blocks) {
-      if (block.analyze_liveliness_phase2(cache->basic_blocks, in.instructions)) {
+    for (auto& block : cache->control_flow.basic_blocks) {
+      if (block.analyze_liveliness_phase2(cache->control_flow.basic_blocks, in.instructions)) {
         changed = true;
       }
     }
   } while (changed);
 
   // phase 3
-  for (auto& block : cache->basic_blocks) {
-    block.analyze_liveliness_phase3(cache->basic_blocks, in.instructions);
+  for (auto& block : cache->control_flow.basic_blocks) {
+    block.analyze_liveliness_phase3(cache->control_flow.basic_blocks, in.instructions);
   }
 
   // phase 4
@@ -192,141 +122,6 @@ void analyze_liveliness(RegAllocCache* cache, const AllocationInput& in) {
       }
     }
   }
-}
-
-void RegAllocBasicBlock::analyze_liveliness_phase1(const std::vector<RegAllocInstr>& instructions) {
-  for (int i = instr_idx.size(); i-- > 0;) {
-    auto ii = instr_idx.at(i);
-    auto& instr = instructions.at(ii);
-    auto& lv = live.at(i);
-    auto& dd = dead.at(i);
-
-    // make all read live out
-    lv.clear();
-    for (auto& x : instr.read) {
-      lv.insert(x.id);
-    }
-
-    // kill things which are overwritten
-    dd.clear();
-    for (auto& x : instr.write) {
-      if (!lv[x.id]) {
-        dd.insert(x.id);
-      }
-    }
-
-    use.bitwise_and_not(dd);
-    use.bitwise_or(lv);
-
-    defs.bitwise_and_not(lv);
-    defs.bitwise_or(dd);
-  }
-}
-
-bool RegAllocBasicBlock::analyze_liveliness_phase2(std::vector<RegAllocBasicBlock>& blocks,
-                                                   const std::vector<RegAllocInstr>& instructions) {
-  (void)instructions;
-  bool changed = false;
-  auto out = defs;
-
-  for (auto s : succ) {
-    out.bitwise_or(blocks.at(s).input);
-  }
-
-  IRegSet in = use;
-  IRegSet temp = out;
-  temp.bitwise_and_not(defs);
-  in.bitwise_or(temp);
-
-  if (in != input || out != output) {
-    changed = true;
-    input = in;
-    output = out;
-  }
-
-  return changed;
-}
-
-void RegAllocBasicBlock::analyze_liveliness_phase3(std::vector<RegAllocBasicBlock>& blocks,
-                                                   const std::vector<RegAllocInstr>& instructions) {
-  (void)instructions;
-  IRegSet live_local;
-  for (auto s : succ) {
-    live_local.bitwise_or(blocks.at(s).input);
-  }
-
-  for (int i = instr_idx.size(); i-- > 0;) {
-    auto& lv = live.at(i);
-    auto& dd = dead.at(i);
-
-    IRegSet new_live = live_local;
-    new_live.bitwise_and_not(dd);
-    new_live.bitwise_or(lv);
-
-    lv = live_local;
-    live_local = new_live;
-  }
-}
-
-std::string RegAllocBasicBlock::print_summary() {
-  std::string result = "block " + std::to_string(idx) + "\nsucc: ";
-  for (auto s : succ) {
-    result += std::to_string(s) + " ";
-  }
-  result += "\npred: ";
-  for (auto p : pred) {
-    result += std::to_string(p) + " ";
-  }
-  result += "\nuse: ";
-  for (int x = 0; x < use.size(); x++) {
-    if (use[x]) {
-      result += std::to_string(x) + " ";
-    }
-  }
-  result += "\ndef: ";
-  for (int x = 0; x < defs.size(); x++) {
-    if (defs[x]) {
-      result += std::to_string(x) + " ";
-    }
-  }
-  result += "\ninput: ";
-  for (int x = 0; x < input.size(); x++) {
-    if (input[x]) {
-      result += std::to_string(x) + " ";
-    }
-  }
-  result += "\noutput: ";
-  for (int x = 0; x < output.size(); x++) {
-    if (output[x]) {
-      result += std::to_string(x) + " ";
-    }
-  }
-
-  return result;
-}
-
-std::string RegAllocBasicBlock::print(const std::vector<RegAllocInstr>& insts) {
-  std::string result = print_summary() + "\n";
-  int k = 0;
-  for (auto instr : instr_idx) {
-    std::string line = insts.at(instr).print();
-    constexpr int pad_len = 30;
-    if (line.length() < pad_len) {
-      // line.insert(line.begin(), pad_len - line.length(), ' ');
-      line.append(pad_len - line.length(), ' ');
-    }
-
-    result += "  " + line + " live: ";
-    for (int x = 0; x < live.at(k).size(); x++) {
-      if (live.at(k)[x]) {
-        result += std::to_string(x) + " ";
-      }
-    }
-    result += "\n";
-
-    k++;
-  }
-  return result;
 }
 
 /*!
@@ -794,6 +589,12 @@ bool try_spill_coloring(int var, RegAllocCache* cache, const AllocationInput& in
 
     if (bonus.load || bonus.store) {
       cache->stack_ops.at(instr).ops.push_back(bonus);
+      if (bonus.load) {
+        cache->stats.num_spill_ops++;
+      }
+      if (bonus.store) {
+        cache->stats.num_spill_ops++;
+      }
     }
   }
   return true;
@@ -909,4 +710,138 @@ bool run_allocator(RegAllocCache* cache, const AllocationInput& in, int debug_tr
     }
   }
   return true;
+}
+
+namespace {
+/*!
+ * Print out the state of the RegAllocCache after doing analysis.
+ */
+void print_analysis(const AllocationInput& in, RegAllocCache* cache) {
+  fmt::print("[RegAlloc] Basic Blocks\n");
+  fmt::print("-----------------------------------------------------------------\n");
+  for (auto& b : cache->control_flow.basic_blocks) {
+    fmt::print("{}\n", b.print(in.instructions));
+  }
+
+  printf("[RegAlloc] Alive Info\n");
+  printf("-----------------------------------------------------------------\n");
+  // align to where we start putting live stuff
+  printf("      %30s    ", "");
+  for (int i = 0; i < cache->max_var; i++) {
+    printf("%2d ", i);
+  }
+  printf("\n");
+  printf("_________________________________________________________________\n");
+  for (uint32_t i = 0; i < in.instructions.size(); i++) {
+    std::vector<bool> ids_live;
+    std::string lives;
+
+    ids_live.resize(cache->max_var, false);
+
+    for (int j = 0; j < cache->max_var; j++) {
+      if (cache->live_ranges.at(j).is_live_at_instr(i)) {
+        ids_live.at(j) = true;
+      }
+    }
+
+    for (uint32_t j = 0; j < ids_live.size(); j++) {
+      if (ids_live[j]) {
+        char buff[256];
+        sprintf(buff, "%2d ", (int)j);
+        lives.append(buff);
+      } else {
+        lives.append(".. ");
+      }
+    }
+
+    if (in.debug_instruction_names.size() == in.instructions.size()) {
+      std::string code_str = in.debug_instruction_names.at(i);
+      if (code_str.length() >= 50) {
+        code_str = code_str.substr(0, 48);
+        code_str.push_back('~');
+      }
+      printf("[%03d] %30s -> %s\n", (int)i, code_str.c_str(), lives.c_str());
+    } else {
+      printf("[%03d] %30s -> %s\n", (int)i, "???", lives.c_str());
+    }
+  }
+}
+}  // namespace
+
+/*!
+ * The top-level register allocation algorithm!
+ */
+AllocationResult allocate_registers(const AllocationInput& input) {
+  AllocationResult result;
+  RegAllocCache cache;
+  cache.is_asm_func = input.is_asm_function;
+
+  // if desired, print input for debugging.
+  if (input.debug_settings.print_input) {
+    print_allocate_input(input);
+  }
+
+  // first step is analysis
+  find_basic_blocks(&cache.control_flow, input);
+  analyze_liveliness(&cache, input);
+  if (input.debug_settings.print_analysis) {
+    print_analysis(input, &cache);
+  }
+
+  // do constraints first, to get them out of the way
+  do_constrained_alloc(&cache, input, input.debug_settings.trace_debug_constraints);
+  // the user may have specified invalid constraints, so we should attempt to find conflicts now
+  // rather than having the register allocation mysteriously fail later on or silently ignore a
+  // constraint.
+  if (!check_constrained_alloc(&cache, input)) {
+    result.ok = false;
+    fmt::print("[RegAlloc Error] Register allocation has failed due to bad constraints.\n");
+    return result;
+  }
+
+  // do the allocations!
+  if (!run_allocator(&cache, input, input.debug_settings.allocate_log_level)) {
+    result.ok = false;
+    fmt::print("[RegAlloc Error] Register allocation has failed.\n");
+    return result;
+  }
+
+  // prepare the result
+  result.ok = true;
+  result.needs_aligned_stack_for_spills = cache.used_stack;
+  result.stack_slots_for_spills = cache.current_stack_slot;
+  result.stack_slots_for_vars = input.stack_slots_for_stack_vars;
+
+  // check for use of saved registers
+  for (auto sr : emitter::gRegInfo.get_all_saved()) {
+    bool uses_sr = false;
+    for (auto& lr : cache.live_ranges) {
+      for (int instr_idx = lr.min; instr_idx <= lr.max; instr_idx++) {
+        if (lr.get(instr_idx).reg == sr) {
+          uses_sr = true;
+          break;
+        }
+      }
+      if (uses_sr) {
+        break;
+      }
+    }
+    if (uses_sr) {
+      result.used_saved_regs.push_back(sr);
+    }
+  }
+  // result.ass_as_ranges = std::move(cache.live_ranges);
+  for (auto& lr : cache.live_ranges) {
+    result.ass_as_ranges.push_back(AssignmentRange(lr.min, lr.is_alive, lr.assignment));
+  }
+  result.stack_ops = std::move(cache.stack_ops);
+
+  // final result print
+  if (input.debug_settings.print_result) {
+    print_result(input, result);
+  }
+
+  result.num_spills = cache.stats.num_spill_ops;
+
+  return result;
 }